@@ -72,7 +72,7 @@ extension GenericDomain {
 /**
  Generic variable for the reader implementation
  */
-protocol GenericVariable: GenericVariableMixable {
+protocol GenericVariable: GenericVariableMixable, Sendable {
     /// The filename of the variable. Typically just `temperature_2m`. Level is used to store mutliple levels or ensemble members in one file
     /// NOTE: `level` has been replaced with `ensembleMemberLevel` in settings
     var omFileName: (file: String, level: Int) { get }
