@@ -11,7 +11,7 @@ extension Sequence {
         }
         return values
     }
-
+    
     func asyncFlatMap<T>(
         _ transform: (Element) async throws -> [T]
     ) async rethrows -> [T] {
@@ -23,13 +23,13 @@ extension Sequence {
         }
         return values
     }
-
+    
     func asyncCompactMap<T>(
         _ transform: (Element) async throws -> T?
     ) async rethrows -> [T] {
         var values = [T]()
         values.reserveCapacity(self.underestimatedCount)
-
+        
         for element in self {
             guard let result = try await transform(element) else {
                 continue
@@ -38,7 +38,9 @@ extension Sequence {
         }
         return values
     }
+}
 
+extension Sequence where Element: Sendable {
     /// Execute a closure for each element concurrently
     /// `nConcurrent` limits the number of concurrent tasks
     func foreachConcurrent(
@@ -56,11 +58,11 @@ extension Sequence {
             try await group.waitForAll()
         }
     }
-
+    
     /// Execute a closure for each element concurrently and return a new value
     /// `nConcurrent` limits the number of concurrent tasks
     /// Note: Results are ordered which may have a performance penalty
-    func mapConcurrent<T>(
+    func mapConcurrent<T: Sendable>(
         nConcurrent: Int,
         body: @escaping @Sendable (Element) async throws -> T
     ) async rethrows -> [T] {
@@ -80,12 +82,14 @@ extension Sequence {
             return results.sorted(by: { $0.0 < $1.0 }).map { $0.1 }
         }
     }
+}
 
+extension Sequence where Element: Sendable, Self: Sendable {
     /// Execute a closure for each element concurrently and return a new value
     /// Returns an `AsyncStream` to process in a pipeline
     /// `nConcurrent` limits the number of concurrent tasks
     /// Note: Results are ordered which may have a performance penalty
-    func mapStream<T>(
+    func mapStream<T: Sendable>(
         nConcurrent: Int,
         body: @escaping @Sendable (Element) async throws -> T
     ) -> AsyncThrowingStream<T, Error> {
@@ -136,12 +140,14 @@ extension AsyncSequence {
         }
         return results
     }
+}
 
+extension AsyncSequence where Element: Sendable, Self: Sendable {
     /// Execute a closure for each element concurrently and return a new value
     /// Returns an `AsyncStream` to process in a pipeline
     /// `nConcurrent` limits the number of concurrent tasks
     /// Note: Results are ordered which may have a performance penalty
-    func mapStream<T>(
+    func mapStream<T: Sendable>(
         nConcurrent: Int,
         body: @escaping @Sendable (Element) async throws -> T
     ) -> AsyncThrowingStream<T, Error> {
