@@ -14,7 +14,7 @@ import OmFileFormat
  */
 struct SatelliteDownloadCommand: AsyncCommand {
     /// 6k locations require around 200 MB memory for a yearly time-series
-    static var nLocationsPerChunk = 6_000
+    static let nLocationsPerChunk = 6_000
 
     struct Signature: CommandSignature {
         @Argument(name: "domain")
